fix cyclic reference
